refactor: modular hero components
