--- components/Hero.module.css
@@ -1,97 +1,0 @@
-.hero {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  --loader-color: #a5a5b0;
-  color: var(--loader-color);
-  height: 100vh;
-  overflow: hidden;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  height: 100%;
-  padding: 2rem;
-  animation: hero-fade 0.8s ease-in-out;
-}
-
-.hero::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(165, 165, 176, 0.5);
-  z-index: 1;
-}
-
-.hero > * {
-  position: relative;
-  z-index: 2;
-}
-
-@keyframes hero-fade {
-  from {
-    opacity: 0;
-    transform: scale(0.95);
-  }
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-@media (max-width: 768px) {
-  .hero {
-    background: url('https://images.unsplash.com/photo-1495446815901-a7297e633e8d?auto=format&fit=crop&w=1280&q=80') center/cover no-repeat;
-  }
-}
-
-
-.hero h1 {
-  font-size: 3rem;
-}
-
-.hero p {
-  font-size: 1.5rem;
-}
-
-@media (max-width: 600px) {
-  .content {
-    padding: 1.5rem;
-  }
-
-  .hero h1 {
-    font-size: 2rem;
-  }
-
-  .hero p {
-    font-size: 1rem;
-  }
-}
-
-@media (min-width: 768px) {
-  .hero h1 {
-    font-size: 4rem;
-  }
-
-  .hero p {
-    font-size: 2rem;
-  }
-}
-
-@media (min-width: 1024px) {
-  .hero h1 {
-    font-size: 5rem;
-  }
-
-  .hero p {
-    font-size: 2.5rem;
-  }
-}
