+++ components/hero/BookbinderyHeros.tsx
@@ -1,0 +1,5 @@
+import HeroHeritageClean from './HeroHeritageClean';
+
+export default function BookbinderyHeros() {
+  return <HeroHeritageClean />;
+}
