+++ components/hero/index.ts
@@ -1,0 +1,4 @@
+export { default } from './BookbinderyHeros';
+export { default as MediaLoop } from './MediaLoop';
+export { default as BookLogo } from './BookLogo';
+export { default as HeroHeritageClean } from './HeroHeritageClean';
